LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/evaluation-order-1-expected.txt

CONSOLE MESSAGE: Error: error

FAIL Test evaluation order of modules assert_array_equals: lengths differ, expected array ["step-1-1", "step-1-2", "microtask", "global-error", "error", "script-load", "global-load"] length 7, got ["step-1-1", "step-1-2", "global-error", "error", "script-load", "global-load"] length 6
